+++ src/EventManager.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+
+namespace PPG
+{
+	class EventManager
+	{
+	public:
+
+		virtual void onObjectEvent(std::string objName, std::string eventName) = 0;
+
+	};
+}
