+++ src/GeneratorHelper.h
@@ -1,0 +1,45 @@
+#pragma once
+
+#include "Relation.h"
+#include "Randomizer.h"
+#include "PuzzGenCore.h"
+#include "Puzzle.h"
+#include "Logger.h"
+#include "Rule.h"
+
+namespace PPG
+{
+
+	using NodePair = Pair<Node*, Node*>;
+	using NodePairVec = Vec<NodePair>;
+	using NodeVec = Vec<Node*>;
+	using RuleVec = Vec<Rule>;
+
+	class GeneratorHelper
+	{
+
+	public:
+		static bool checkCreatesCircularDependency(NodePair P, Relation* R);
+		static bool checkCreatesExclusiveDependency(NodePair P, Relation* R);
+		static bool findNodeSequential(Node* N, NodePair start, Relation* R);
+		static bool checkEquality(Node* N1, Node* N2);
+		static bool checkMetaEqualOccurance(NodePair P, Relation* R);
+
+
+		static bool checkCompatibilityBasicRules(Node* S, Node* N, Relation* R);
+		static bool checkCompatibilityCustomRules(NodeVec nodes, Node* S, Node* N, Relation* R, RuleVec rules);
+
+		static bool checkCompatibilityRuleTypeAfter(NodeVec nodes, Node* S, Node* N, Relation* R, Rule rule, bool isStrict);
+		static bool checkCompatibilityRuleTypeBefore(NodeVec nodes, Node* S, Node* N, Relation* R, Rule rule, bool isStrict);
+
+		static NodeVec filterCompatibleNodes(Node* N, Relation* R, NodeVec nodes, RuleVec rules);
+
+		static bool checkMetaEqualOccuranceByNode(Node* N, Relation* R);
+
+		static bool isRuleNodeEqual(Node* N, Object* ruleObject, State* ruleState);
+		static bool isRuleObjectEqual(Object* o1, Object* o2);
+		static bool isRuleStateEqual(State* s1, State* s2);
+
+	};
+
+}
