+++ src/GraphNode.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include "PuzzGenCore.h"
+
+namespace PPG
+{
+	class GraphNode
+	{
+	public:
+		Object* getObject();
+		State* getState();
+		Vec<GraphNode*> getChildren();
+
+		void setObject(Object* object);
+		void setState(State* state);
+		void setChildren(Vec<GraphNode*> children);
+		void addChild(GraphNode* child);
+
+	private:
+		Object* object;
+		State* state;
+		Vec<GraphNode*> children;
+
+	};
+}
+
+
