+++ src/ObjectMetaData.h
@@ -1,0 +1,12 @@
+#pragma once
+
+/**
+* Dataholder Interface Class for Metadata of PuzzleObjects
+* For Developers who want to add metadata to their PuzzleObjects
+* Use the MetaData to help with interpretation
+*
+* This class can be set to PuzzleObjects and read for easier interpretation of those objects 
+*/
+
+class ObjectMetaData {};
+
