+++ src/StateTransition.cpp
@@ -8,11 +8,11 @@ namespace PPG
 		return &this->transitionMap;
 	}
 
-	void StateTransition::addTransition(std::string eventName, PuzzleState origState, PuzzleState newState)
+	void StateTransition::addTransition(std::string eventName, State origState, State newState)
 	{
 
 		bool eventEntryFound = (this->transitionMap.find(eventName) != this->transitionMap.end());
-		std::pair<PuzzleState, PuzzleState> newStatePair;
+		std::pair<State, State> newStatePair;
 		newStatePair.first = origState;
 		newStatePair.second = newState;
 
@@ -20,7 +20,7 @@ namespace PPG
 			this->transitionMap.find(eventName)->second.push_back(newStatePair);
 		}
 		else {
-			std::vector<std::pair<PuzzleState, PuzzleState>> newVector;
+			std::vector<std::pair<State, State>> newVector;
 			newVector.push_back(newStatePair);
 			this->transitionMap.insert(std::make_pair(eventName, newVector));
 		}
@@ -32,7 +32,7 @@ namespace PPG
 		std::string out = "";
 		for (EVENT_MAP::iterator it = this->transitionMap.begin(); it != this->transitionMap.end(); ++it) {
 			out += "> " + it->first + ": ";
-			for (std::vector<std::pair<PuzzleState, PuzzleState>>::iterator vecIt = it->second.begin(); vecIt != it->second.end(); ++vecIt) {
+			for (std::vector<std::pair<State, State>>::iterator vecIt = it->second.begin(); vecIt != it->second.end(); ++vecIt) {
 				out += "<" + vecIt->first.getStateName() + " -> " + vecIt->second.getStateName() + "> ";
 			}
 			out += "\n";
@@ -41,25 +41,25 @@ namespace PPG
 		return out;
 	}
 
-	std::vector<PuzzleState> StateTransition::getReachableStates()
+	std::vector<State> StateTransition::getReachableStates()
 	{
-		std::vector<PuzzleState> reachables;
+		std::vector<State> reachables;
 		for (EVENT_MAP::iterator it = this->transitionMap.begin(); it != this->transitionMap.end(); ++it) {
-			for (std::vector<std::pair<PuzzleState, PuzzleState>>::iterator vecIt = it->second.begin(); vecIt != it->second.end(); ++vecIt) {
+			for (std::vector<std::pair<State, State>>::iterator vecIt = it->second.begin(); vecIt != it->second.end(); ++vecIt) {
 				reachables.push_back(vecIt->second);
 			}
 		}
 		return reachables;
 	}
 
-	std::vector<std::pair<PuzzleState, PuzzleState>> StateTransition::findTransitions(const std::string name)
+	std::vector<std::pair<State, State>> StateTransition::findTransitions(const std::string name)
 	{
 		EVENT_MAP::iterator it = this->transitionMap.find(name);
 		if (it != this->transitionMap.end()) {
 			return it->second;
 		}
 
-		std::vector<std::pair<PuzzleState, PuzzleState>> a;
+		std::vector<std::pair<State, State>> a;
 		return a;
 	}
 }
